Accumulate the Java hash in an unsigned type
The hash multiplied a signed 32-bit accumulator by 31 each round.  Java defines
that arithmetic to wrap, but in C++ the overflow is undefined, and a string of
43 characters was enough to reach it.

Accumulate in uint32_t and convert once on return.  The values still match
Java for the empty string, "a", "ab", "Hello", "hello", and a long string that
wraps.

+++ link/.local/include/java_hashCode.hpp
@@ -19,7 +19,7 @@
 constexpr int32_t
 java_hashCode(const std::string_view s)
 {
-    int32_t result = 0;
+    uint32_t result = 0;
 
     for (const auto c : s)
     {
@@ -27,5 +27,5 @@ java_hashCode(const std::string_view s)
         result = 31 * result + octet;
     }
 
-    return result;
+    return static_cast<int32_t>(result);
 }
